+++ services/user-svc/src/main/java/edu/utn/frsf/isi/dan/user/dao/PropietarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PropietarioRepository extends JpaRepository<Propietario, Long> {
+public interface PropietarioRepository extends JpaRepository<Propietario, Integer> {
 }
