+++ services/user-svc/src/main/java/edu/utn/frsf/isi/dan/user/model/Propietario.java
@@ -17,7 +17,7 @@ public class Propietario extends Usuario {
     private CuentaBancaria cuentaBancaria;
 
     @Column(name = "hotel_id")
-    private Long idHotel;
+    private Integer idHotel;
     // solo guardo el id del hotel, no la entidad Hotel 
     // porque se gestiona en otro microservicio
     // entonces la consistencia aquí será eventual
